requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs @ file:///croot/attrs_1695717823297/work
backcall==0.2.0
bcrypt==4.2.0
boto3==1.35.10
botocore==1.35.10
braceexpand==0.1.7
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==4.2.4
certifi @ file:///croot/certifi_1720453481653/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1723018376978/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.18.1
cloudpickle==3.0.0
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cryptography==43.0.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
datasets==3.1.0
decorator==4.4.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dtw==1.4.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.4
fastapi==0.115.12
fastdtw==0.3.4
filelock==3.15.4
fire==0.7.0
fonttools @ file:///croot/fonttools_1713551344105/work
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.2.3
gast==0.3.3
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///croot/gitpython_1720455034730/work
google-ai-generativelanguage==0.6.5
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
-e git+https://github.com/jstmn/generative-ai-python-3.8.git@e4257397b4bdb7326962daa1e0f93b0652b08300#egg=google_generativeai
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.66.1
grpcio-status==1.62.3
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1633502707516/work
h5py==3.11.0
-e git+ssh://git@github.com/facebookresearch/habitat-lab.git@d6ed1c0a0e786f16f261de2beafe347f4186d0d8#egg=habitat
habitat-sim==0.1.7
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.6
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1724450538981/work
ifcfg==0.24
imageio @ file:///croot/imageio_1707247282708/work
imageio-ffmpeg @ file:///home/conda/feedstock_root/build_artifacts/imageio-ffmpeg_1717461632069/work
importlib_metadata==8.4.0
importlib_resources @ file:///croot/importlib_resources-suite_1720641103994/work
interegular==0.3.3
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Keras-Preprocessing==1.1.2
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lark==1.2.2
llvmlite==0.41.1
lmdb==1.5.1
lmdeploy==0.7.2.post1
magnum @ file:///opt/conda/envs/py36/conda-bld/habitat-sim_1611971337805/work/build/deps/magnum-bindings/src/python
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib @ file:///croot/matplotlib-suite_1693812469450/work
matplotlib-inline==0.1.7
mdurl==0.1.2
mmengine-lite==0.10.7
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy==0.4.7.1
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.1
numba==0.58.1
numpy==1.22.3
numpy-quaternion @ file:///home/conda/feedstock_root/build_artifacts/quaternion_1676352986710/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.42.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
outlines==0.0.46
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.0.3
paramiko==3.4.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.13.2
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1721059439630/work
pkgutil_resolve_name==1.3.10
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
ply==3.11
pooch @ file:///croot/pooch_1695850093751/work
proglog==0.1.10
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.26.1
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyairports==2.1.1
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycountry==24.6.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyNaCl==1.5.0
pynvml==11.5.3
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytorch-transformers==1.2.0
pytz==2025.2
PyYAML==6.0.2
-e git+https://github.com/xinyu1205/recognize-anything.git@88c2b0ca13e38cca6655f83ad0185271167dbcbf#egg=ram
ray==2.10.0
referencing==0.35.1
regex==2024.7.24
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.20.1
rsa==4.9
s3transfer==0.10.2
sacremoses==0.1.1
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
shortuuid==1.0.13
simplejson==3.19.3
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap @ file:///tmp/build/80754af9/smmap_1611517644793/work
sniffio==1.3.1
stack-data==0.6.3
starlette==0.44.0
sympy==1.13.2
tb-nightly==2.13.0a20230317
tenacity==9.0.0
tensorboard==2.2.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==0.4.12
tokenizers==0.19.1
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==2.4.1
torchaudio==2.4.0
torchvision==0.19.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm @ file:///croot/tqdm_1724853939799/work
traitlets==5.14.3
transformers==4.44.0
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.2
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.33.0
wcwidth==0.2.13
webdataset==0.1.40
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.15.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1724730934107/work
zstandard==0.22.0
